healthcare-poc/backend/requirements.txt
strands-agents==1.20.0
strands-agents-tools==0.2.0
boto3>=1.42.0
python-dotenv==1.0.1
mcp==1.21.0
python-multipart==0.0.9
